made a better build
