+++ dev/lib/publisher.js
@@ -12,24 +12,6 @@ function readManifest(packagePath) {
     return JSON.parse(manifest);
 }
 
-function compress(zipPath) {
-    console.log("zipping...");
-    // zip it.
-    var output = FileSystem.createWriteStream(path.join(zipPath));
-    var archive = archiver("zip", {
-        zlib: { level: 9 } // Sets the compression level.
-    });
-    // listen for all archive data to be written
-    output.on("close", function() {
-        console.log(archive.pointer() + " total bytes");
-        console.log(
-            "archiver has been finalized and the output file descriptor has closed."
-        );
-    });
-    archive.pipe(output);
-    archive.directory(path.join(__dirname, "/build/"), false);
-    archive.finalize();
-}
 const store = function(params) {
     const { manifestPath, tokens, archive } = params;
     const { appId, clientId, clientSecret, refreshToken } = tokens;
@@ -50,34 +32,11 @@ const store = function(params) {
             const body = JSON.parse(result[1]);
             if (body.crxVersion) {
                 console.log("Current version on store: '%s'", body.crxVersion);
-            }
-
-            if (!extensionVersion.isValid(body.crxVersion)) {
-                throw new Error(
-                    sprintf(
-                        "Existing item version '%s' is not a valid Chrome extension version",
-                        body.crxVersion
-                    )
+                console.log(
+                    "Current version on local: '%s'",
+                    appManifest.version
                 );
             }
-            if (!extensionVersion.isValid(appManifest.version)) {
-                throw new Error(
-                    sprintf(
-                        "Version in manifest '%s' is not a valid Chrome extension version",
-                        appManifest.version
-                    )
-                );
-            }
-
-            const isManifestVersionNewer = extensionVersion.lessThan(
-                body.crxVersion,
-                appManifest.version
-            );
-            console.log(
-                body.crxVersion,
-                appManifest.version,
-                isManifestVersionNewer
-            );
             return chromeWebStore.uploadPackage(
                 archive,
                 appId,
