--- dev/lib/readManifest.js
@@ -1,8 +1,0 @@
-const FileSystem = require("fs");
-
-function readManifest(packagePath) {
-    let manifest = fs.readFileSync(packagePath, "utf8");
-    return JSON.parse(manifest);
-}
-
-module.exports = readManifest;
