+++ webpack.config.js
@@ -1,12 +1,11 @@
-var CopyWebpackPlugin = require("copy-webpack-plugin");
 var webpack = require("webpack");
 var SassPlugin = require("sass-webpack-plugin");
 var StringReplacePlugin = require("string-replace-webpack-plugin");
-var FileSystem = require("fs");
 var path = require("path");
-var archiver = require("archiver");
 var publisher = require("./dev/lib/publisher");
 var tokens = require("./tokens");
+var copy = require("./webpack/copy");
+var pack = require("./webpack/pack");
 
 module.exports = {
     entry: {
@@ -23,82 +22,24 @@ module.exports = {
     },
     plugins: [
         new StringReplacePlugin(),
-        // new webpack.ContextReplacementPlugin(
-        //     /\.\/locale$/,
-        //     "empty-module",
-        //     false,
-        //     /js$/
-        // ),
         new webpack.optimize.UglifyJsPlugin({
             compress: { warnings: false },
             output: { comments: false },
             sourceMap: true
         }),
-
+        copy(),
         //new SassPlugin("dev/css/themes/dark.scss"),
-        new CopyWebpackPlugin(
-            [
-                // {output}/file.txt
-                { context: "dev", from: "*", to: "../build/" },
-                { context: "dev", from: "public/**/*", to: "../build/" },
-                {
-                    context: "dev",
-                    from: "js/background.js",
-                    to: "../build/js"
-                },
-                { context: "dev", from: "js/content.js", to: "../build/js" }
-            ],
-            {
-                ignore: ["webpack.config.js", "package.json"]
+        pack({
+            src: path.join(__dirname, "build/"),
+            target: path.join(__dirname, "/linkcast.zip"),
+            callback: params => {
+                publisher.publish({
+                    archive: params.target,
+                    tokens: tokens,
+                    manifestPath: path.join(__dirname, "/dev/manifest.json")
+                });
             }
-        ),
-        function() {
-            this.plugin("done", function(statsData) {
-                var stats = statsData.toJson();
-                var isDev =
-                    process.argv
-                        .filter(arg => {
-                            return arg == "-p";
-                        })
-                        .join() != "-p";
-                if (!stats.errors.length && !isDev) {
-                    var htmlFileName = "dev/popup.html";
-                    var html = FileSystem.readFileSync(
-                        path.join(__dirname, htmlFileName),
-                        "utf8"
-                    );
-                    var htmlOutput = html
-                        .replace("http://localhost:3000/build", "")
-                        .replace("development_mode", "prod");
-                    FileSystem.writeFileSync(
-                        path.join(__dirname, "build", "popup.html"),
-                        htmlOutput
-                    );
-                    console.log("zipping...");
-                    // zip it.
-                    var output = FileSystem.createWriteStream(
-                        path.join(__dirname, "/linkcast.zip")
-                    );
-                    var archive = archiver("zip");
-                    // listen for all archive data to be written
-                    output.on("close", function() {
-                        console.log(archive.pointer() + " total bytes");
-                        console.log(
-                            "archiver has been finalized and the output file descriptor has closed."
-                        );
-                    });
-                    archive.pipe(output);
-                    archive.directory(path.join(__dirname, "/build/"), false);
-                    archive.finalize();
-
-                    publisher.publish({
-                        archive: path.join(__dirname, "/linkcast.zip"),
-                        tokens: tokens,
-                        manifestPath: path.join(__dirname, "/dev/manifest.json")
-                    });
-                }
-            });
-        }
+        })
     ],
     module: {
         rules: [
